use properties class
