+++ build.gradle
@@ -22,7 +22,7 @@ repositories {
 }
 
 group 'bio.terra'
-version '0.0.25-SNAPSHOT' // local change only
+version '0.0.27-SNAPSHOT' // local change only
 sourceCompatibility = JavaVersion.VERSION_1_8
 
 def OPENCENSUS_VERSION = '0.28.+'
