+++ src/main/java/bio/terra/common/stairway/StairwayProperties.java
@@ -1,0 +1,53 @@
+package bio.terra.common.stairway;
+
+import java.time.Duration;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "terra.common.stairway")
+public class StairwayProperties {
+  private boolean forceCleanStart;
+  private boolean migrateUpgrade;
+  private int maxParallelFlights;
+  private Duration quietDownTimeout;
+  private Duration terminateTimeout;
+
+  public boolean isForceCleanStart() {
+    return forceCleanStart;
+  }
+
+  public boolean isMigrateUpgrade() {
+    return migrateUpgrade;
+  }
+
+  public int getMaxParallelFlights() {
+    return maxParallelFlights;
+  }
+
+  public Duration getQuietDownTimeout() {
+    return quietDownTimeout;
+  }
+
+  public Duration getTerminateTimeout() {
+    return terminateTimeout;
+  }
+
+  public void setForceCleanStart(boolean forceCleanStart) {
+    this.forceCleanStart = forceCleanStart;
+  }
+
+  public void setMigrateUpgrade(boolean migrateUpgrade) {
+    this.migrateUpgrade = migrateUpgrade;
+  }
+
+  public void setMaxParallelFlights(int maxParallelFlights) {
+    this.maxParallelFlights = maxParallelFlights;
+  }
+
+  public void setQuietDownTimeout(Duration quietDownTimeout) {
+    this.quietDownTimeout = quietDownTimeout;
+  }
+
+  public void setTerminateTimeout(Duration terminateTimeout) {
+    this.terminateTimeout = terminateTimeout;
+  }
+}
